[2/4] Logical operators for safety_checker results
The operators &&, ||, &= and |= can now be used between two
safety_checkert::resultt objects. They return the "best" (|| and |=) or
"worst" (&& and &=) of two results, and the later two operators
additionally assign that result to the LHS. The assignment operators are
useful when repeatedly performing a safety check and wanting to return a
single error code that either describes the best or worst outcome of all
the safety checks combined.