Abstract data structure for save-points
There is now an interface for a data structure that holds save-points,
rather than just a simple list. The abstract data structure currently
has one concrete subtype, which is a FIFO (with the same behaviour as
the previous list implementation), but this data type will be used to
support more advanced path-resumption strategies.