[path explore 6/7] Reverse goto-priority flag
Provide a new flag, --goto-priority, which specifies whether to first
symbolically execute the target or the next instruction of a goto. The
default goto-priority for bounded model checking remains to execute the
next instruction, but the default for path-exploration is to execute the
target of the jump.

This is useful for specifying whether path exploration should prefer to
make forward progress, or to spend longer executing loops up to the
unwind limit before moving forward.