Symex paths saved onto abstract data structure
This commit introduces a new data type, path_queuet, representing saved
symbolic execution paths, used instead of the std::list. This allows us
to implement custom logic for saving and removing paths, rather than
simply pushing to the back and popping the front.

This commit introduces a concrete instantiation of path_queuet, called
path_fifot, with the same functionality as the old std::list. This is in
preparation for further instantiations that implement more sophisticated
logic for deciding which path should be resumed next.