Add --paths tests for exploration strategies
This commit introduces a regression test suite specifically for testing
path exploration strategies when using cbmc with --paths. This test
suite is intended to ensure that paths are pushed and popped onto the
workqueue in the correct order as specified by their search strategy.
This commit contains a test suite for the FIFO exploration strategy.

This commit introduces a script for generating test.desc files. This is
needed because the multi-line regexes for matching the order of path
pushes and pops are very cumbersome to write and difficult to read.
Thus, the test.desc files are generated from a test.in file, which
includes a human-friendly description of what push, pop and execute
events should be expected during symbolic execution. A new script,
scripts/make_descs.py, interprets those events and generates a test.desc
from them before running the test suite.