Add progressive path exploration strategy
Path exploration can now be run with a new strategy: "progressive-fifo".
This strategy prefers to pop the jumps of a goto instruction rather than
the next instruction. This is intended to make forward progress through
a program rather than dwelling on loops. This commit adds tests for this
strategy also.

A new class is introduced to hold the names, descriptions and
constructor thunks for all path exploration strategies. This provides a
uniform way for strategy descriptions to be displayed, which can be done
with the --show-symex-strategies flag.