Path_storaget has reference to goto_modelt
Derived classes of path_storaget must now be constructed with a
reference to a goto_model. Clients who don't have a goto_model, but
are sure that they don't intend to use the path_storaget anyway, can
construct a degenerate_path_storaget instead which supplies its own
goto_model and enforces that it not be used.

This is in preparation for path exploration strategies that make use of
the goto-program to decide what path to execute next.