Computation of lines reachable from lexical blocks
CBMC can now compute which lines can be reached upon executing a lexical
block (not a basic block, i.e. this takes into account arbitrarily
nested blocks). The lines transitively reached from the beginning of a
block include all lines nested within that block, as well as all lines
in all of the functions that are called from that block.

The result of this computation can be displayed using the new flag
--show-transitive-blocks. This flag is used by several regression tests,
which demonstrate what lines should and shouldn't be reached from which
blocks.