Add degenerate_path_storaget
This class is used as a placeholder by clients who will never wish to
perform path exploration. This is in preparation for commits that will
add new parameters to the constructor for path_storaget and its
subclasses, and which are inconvenient for clients that do not wish to
to path exploration to supply.