Add hybrid BDFS path exploration strategy
From the help text: explore breadth-first until we have N saved paths,
at which point explore depth-first from each of those paths in
round-robin order. N defaults to 8 and can be specified with the flag
--bdfs-diameter N.

This strategy is intended to be a nice compromise between low RAM usage
and high coverage. The fact that we end paths as quickly as possible
keeps memory use low, while branching very high up in the exploration
tree means that we explore many different parts of the program.