Add merge-at-end DFS strategy
This strategy explores a single path. When it has finished, instead of
popping the most-recently pushed path (as DFS would do), it instead pops
N paths and enables path merging. In other words, as soon as it reaches
the leaf of a tree, it does path merging for a subtree that includes
that leaf and that is N control points high. The value of N is set by
using the --end-merge-depth flag.

This is intended to deal with codebases that have a large teardown bit
of code that all paths end up on, and where symbolic execution spends a
long time fruitlessly stuck in that teardown code, when we really want
to direct it to exploring other paths further up the tree.