--stop-on-fail now stops on failed path
When doing path exploration, one sometimes wants CBMC to halt symbolic
execution as soon as it encounters a single bug, rather than continuing
to explore the program for all bugs. This behaviour can now be enabled
by passing the --stop-on-fail flag together with the --paths flag.

Prior to this commit, the --stop-on-fail flag --- which was written with
model-checking rather than path exploration in mind --- would explore
all paths, but only print out a single failed property, which isn't
terribly useful because it uses just as much time as printing out all
properties. This commit makes this flag more useful when used in
conjunction with --paths, while not changing its behaviour for model
checking.