Add clear() to path exploration worklist
Subclasses of path_storaget are now required to implement a clear()
method. This is so that if we terminate path exploration early, the
worklist can be cleared.

This commit facilitates writing tests that ensure that the worklist is
empty iff symbolic execution has finished; this should happen naturally
if symex runs its course over the entire program, but might not happen
if we terminate symex early.

This commit is in preparation for a future commit that makes the
--stop-on-fail flag terminate symbolic execution early.